\name{Alev Baysoy}
\title{SeqAlign}
\usage{
Align two sequences with one another using the Smith-Waterman algorithm with an affine gap penalty
}
\description{
Prints out score matrix, alignment maximum score, and optimal sequence alignment for two sequences
}
\examples{
runSW(inputFile, scoreFile, openGap, extGap)
}
